[async.forEach] Cleanup implementation.
No longer uses callbacks, but rather uses the Twig.Promise
implementation.
